Set locale to C.UTF-8 (#908)
If the locale environment variables are not set to `C.UTF-8` and
`LOCALE_ARCHIVE` is not set to a suitable path, then the program fails
with an invalid character error when a non-ASCII character is printed.

With this commit, locale environment variables are checked and set to
`C.UTF-8` when necessary.  When handing off, any changes are propagated.
With `--no-handoff`, the program must be re-executed in order for the
new settings to take effect.  Note that care must be taken to not print
any non-ASCII characters before handoff/re-execution.

This commit is experimental.  This hack is not suitable if locale
support is required.